fix: Error in function type

+++ grupo2Backend/src/main/java/com/grupo2backend/services/OrdenService.java
@@ -33,7 +33,7 @@ public class OrdenService {
         return ordenRepository.findAll();
     }
 
-    public Optional<Long> getOrdenProcesoByIdCliente(Long idCliente){
+    public Optional<Integer> getOrdenProcesoByIdCliente(Long idCliente){
               return ordenRepository.findByEstadoAndIdCliente(idCliente);
     }
